price - change from float to double

+++ src/main/java/classes/model/Item.java
@@ -6,5 +6,5 @@ import jdk.nashorn.internal.ir.annotations.Immutable;
 public class Item {
     public String name;
     public int quantity;
-    public float price;
+    public Double price;
 }
